requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
nltk==3.8.1
textblob==0.17.1
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
websockets==12.0
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
yfinance==0.2.28
praw==7.7.1
tweepy==4.14.0
newsapi-python==0.2.7
python-dotenv==1.0.0
schedule==1.2.1
logging==0.4.9.6
dataclasses==0.6
typing-extensions==4.9.0
scipy==1.11.4
ta==0.10.2
python-dateutil==2.8.2
pytz==2023.3
supabase==2.3.4
psycopg2-binary==2.9.9